requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
flask-restplus==0.13.0
flask-restx==1.3.0
grpcio==1.66.1
grpcio-tools==1.66.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
protobuf==5.28.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
swagger-ui==0.1.2
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
